Test dsa with nox509

+++ t/018-DigestSignatureAlgorithms.t
@@ -1,18 +1,50 @@
 use strict;
 use warnings;
 
-use Test::More tests => 1044;
+use Test::More tests => 1224;
 use XML::Sig;
 use File::Which;
 
-# DSA Keys with X509
 my @hash = qw/sha1 sha224 sha256 sha384 sha512 ripemd160/;
 foreach my $sigalg (@hash) {
+    # DSA Keys with noX509
+    foreach my $digalg (@hash) {
+        my $sig = XML::Sig->new( {
+            digest_hash => $digalg,
+            sig_hash    => $sigalg,
+            x509        => 0,
+            key         => 't/dsa.private.key',
+        } );
+        isa_ok( $sig, 'XML::Sig' );
+
+        my $signed = $sig->sign('<foo ID="123"></foo>');
+        ok($signed, "XML Signed Sucessfully using dsa-$sigalg digest: $digalg");
+
+        $sig = XML::Sig->new( );
+        my $is_valid = $sig->verify( $signed );
+        ok( $is_valid == 1, "XML::Sig signed Validated using X509Certificate");
+
+        SKIP: {
+            skip "xmlsec1 not installed", 2 unless which('xmlsec1');
+
+            ok( (open XML, '>', "t/tmp-dsa-$sigalg-nox509-$digalg.xml"), "File t/tmp-dsa-$sigalg-nox509-$digalg.xml opened for write");
+            print XML $signed;
+            close XML;
+
+            my $verify_response = `xmlsec1 --verify --id-attr:ID "foo" t/tmp-dsa-$sigalg-nox509-$digalg.xml 2>&1`;
+            ok( $verify_response =~ m/^OK/, "t/tmp-dsa-$sigalg-nox509-$digalg.xml is verified using xmlsec1" )
+                or warn "calling xmlsec1 failed: '$verify_response'\n";
+            unlink "t/tmp-dsa-$sigalg-nox509-$digalg.xml";
+        }
+    }
+
+    # DSA Keys with noX509
     foreach my $digalg (@hash) {
         my $sig = XML::Sig->new( {
             digest_hash => $digalg,
             sig_hash    => $sigalg,
             x509        => 1,
+            cert        => 't/dsa.public.pem',
             key         => 't/dsa.private.key',
         } );
         isa_ok( $sig, 'XML::Sig' );
@@ -31,10 +63,15 @@ foreach my $sigalg (@hash) {
             print XML $signed;
             close XML;
 
-            my $verify_response = `xmlsec1 --verify --id-attr:ID "foo" t/tmp-dsa-$sigalg-x509-$digalg.xml 2>&1`;
+            my $verify_response = `xmlsec1 --verify --id-attr:ID "foo" --pubkey-cert-pem t/dsa.public.pem --trusted-pem t/dsa.public.pem t/tmp-dsa-$sigalg-x509-$digalg.xml 2>&1`;
             ok( $verify_response =~ m/^OK/, "t/tmp-dsa-$sigalg-x509-$digalg.xml is verified using xmlsec1" )
                 or warn "calling xmlsec1 failed: '$verify_response'\n";
-            unlink "t/tmp-dsa-$sigalg-x509-$digalg.xml";
+            if ($verify_response =~ m/^OK/) {
+                unlink "t/tmp-dsa-$sigalg-x509-$digalg.xml";
+            } else{
+                print $signed;
+                die;
+            }
         }
     }
 
@@ -155,6 +192,7 @@ foreach my $sigalg (@hash) {
                     digest_hash => $digalg,
                     sig_hash    => $sigalg,
                     key => 't/ecdsa.private.pem',
+                    x509 => 0,
                     }
                 );
         isa_ok( $sig, 'XML::Sig' );
